+++ CHANGELOG.md
@@ -17,10 +17,13 @@
   </summary>
 
 ### :bug: Bug Fix
+- `common`
+  - [#1415](https://github.com/wix-incubator/rich-content/pull/1415) tooltip setstate on an unmounted component and ssr hydration fixed 
 - `editor`
   - [#1410](https://github.com/wix-incubator/rich-content/pull/1410) fix space deletes atomic block by disabling keyboard inputs on atomic blocks
-- `editor`
   - [#1381](https://github.com/wix-incubator/rich-content/pull/1381) fix blur editor on esc click
+
+
 </details>
 <hr/>